Fix camera info topic subscription in image panel
The image panel had incorrect logic for the camera info subscription. It was checking the topic against supported image schemas rather than calibration schemas.

This change fixes the logic by adding the list of supported calibration schemas.
