+++ packages/studio-base/src/panels/Image/hooks/index.ts
@@ -2,4 +2,4 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
-export { useImagePanelMessages, ANNOTATION_DATATYPES } from "./useImagePanelMessages";
+export { useImagePanelMessages } from "./useImagePanelMessages";
